refactor(enhancement): extract model and data hooks
